[CONFIGURATION] * BUG: Libraries and include files are automatically searched inside   64 bits specific directories.
[DEVELOPMENT ENVIRONMENT]
* NEW: Automatically convert old Border property to the new Border and
  Resizable properties.
* BUG: The "Theme" tab of the option dialog is correctly initialized the 
  first time Gambas is run.
* BUG: Going to a non-existing line number in the editor does not crash the
  IDE anymore.
* NEW: The IDE forms now use the new Border, Resizable and Type properties
  of the Window class.

[COMPILER]
* BUG: Do not print error messages on unexpected strings with newline 
  inside. Just claim that an unexpected string has been encountered.

[GB.EVAL]
* BUG: Do not print error messages on unexpected strings with newline 
  inside. Just claim that an unexpected string has been encountered.

[GB.GTK]
* NEW: The old Border property of the Window class has been replaced by a 
  two boolean properties: Border and Resizable.
* NEW: Window.Type is a new property to set the window manager type of a 
  window: Normal, ToolBar, Splash...
* NEW: The Window.ToolBox property has been removed.
* NEW: A newly opened non-modal window stays automatically on top of the
  application main window.

[GB.QT]
* NEW: The old Border property of the Window class has been replaced by a 
  two boolean properties: Border and Resizable.
* NEW: Window.Type is a new property to set the window manager type of a 
  window: Normal, ToolBar, Splash...
* NEW: The Window.ToolBox property has been removed.
* NEW: A newly opened non-modal window stays automatically on top of the
  application main window.


git-svn-id: svn://localhost/gambas/trunk@1321 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ gb.gtk/src/gcontrol.cpp
@@ -756,11 +756,11 @@ gMainWindow* gControl::topLevel()
 	return (gMainWindow *)child;
 }
 
-long gControl::handle()
+int gControl::handle()
 {
 	#ifndef GAMBAS_DIRECTFB
 	#ifdef GDK_WINDOWING_X11
-	return (long)GDK_WINDOW_XID(border->window);
+	return GDK_WINDOW_XID(border->window);
 	#else
 	stub("no-X11/gControl::handle()");
 	return 0;
